+++ govlearn-api/src/test/java/com/unimuenster/govlearnapi/group/controller/content/DeleteContentForAllMembers.java
@@ -1,6 +1,7 @@
-package com.unimuenster.govlearnapi.group.controller;
+package com.unimuenster.govlearnapi.group.controller.content;
 
 import com.unimuenster.govlearnapi.common.responsewrapper.Response;
+import com.unimuenster.govlearnapi.group.controller.GroupTestBase;
 import com.unimuenster.govlearnapi.group.controller.wsto.DeleteContentForGroupWsTo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
