+++ govlearn-api/src/test/java/com/unimuenster/govlearnapi/group/controller/content/GroupContentControllerTest.java
@@ -1,6 +1,7 @@
-package com.unimuenster.govlearnapi.group.controller;
+package com.unimuenster.govlearnapi.group.controller.content;
 
-import com.unimuenster.govlearnapi.group.controller.wsto.AddContentToGroupWsTo;
+import com.unimuenster.govlearnapi.group.controller.GroupTestBase;
+import com.unimuenster.govlearnapi.group.controller.wsto.AddContentToMemberWsTo;
 import com.unimuenster.govlearnapi.group.entity.Group;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -10,8 +11,7 @@ import org.springframework.http.ResponseEntity;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-public class AddContentToAllMembersTest extends GroupTestBase {
-
+public class GroupContentControllerTest extends GroupTestBase {
     @BeforeEach
     void setUp() {
         setCurrentUser(initializerService.getUser2());
@@ -19,20 +19,19 @@ public class AddContentToAllMembersTest extends GroupTestBase {
         addMember();
     }
 
-
     @Test
-    void addContentToAll() {
+    void addContentTest() {
 
         Long courseId = 1L;
 
-        AddContentToGroupWsTo addContentToGroupWsTo = AddContentToGroupWsTo
+        AddContentToMemberWsTo addMemberWsTo = AddContentToMemberWsTo
                 .builder()
-                .groupId(getGroup().getId())
+                .memberId(currentMember.getId())
                 .courseId(courseId)
                 .build();
 
 
-        ResponseEntity responseEntity = groupContentController.addContentToAll(addContentToGroupWsTo);
+        ResponseEntity responseEntity = groupContentController.addContent(addMemberWsTo);
 
         Optional<Group> byId = groupRepository.findById(getGroup().getId());
 
@@ -41,21 +40,20 @@ public class AddContentToAllMembersTest extends GroupTestBase {
         assertEquals(1, byId.get().getMembers().get(0).getCourses().size());
         assertEquals(courseId, byId.get().getMembers().get(0).getCourses().get(0).getId());
     }
-
     @Test
     void addContentButNotAdmin() {
 
         setCurrentUser(initializerService.getUser1());
 
-        AddContentToGroupWsTo addContentToGroupWsTo = AddContentToGroupWsTo
+        AddContentToMemberWsTo addMemberWsTo = AddContentToMemberWsTo
                 .builder()
-                .groupId(getGroup().getId())
+                .memberId(currentMember.getId())
                 .courseId(1L)
                 .build();
 
-
-        ResponseEntity responseEntity = groupContentController.addContentToAll(addContentToGroupWsTo);
+        ResponseEntity responseEntity = groupContentController.addContent(addMemberWsTo);
 
         assertEquals(HttpStatus.FORBIDDEN, responseEntity.getStatusCode());
+
     }
 }
