+++ govlearn-api/src/test/java/com/unimuenster/govlearnapi/group/controller/member/GetGroupsByUserTest.java
@@ -1,6 +1,7 @@
-package com.unimuenster.govlearnapi.group.controller;
+package com.unimuenster.govlearnapi.group.controller.member;
 
 import com.unimuenster.govlearnapi.common.responsewrapper.Response;
+import com.unimuenster.govlearnapi.group.controller.GroupTestBase;
 import com.unimuenster.govlearnapi.group.controller.wsto.AddMemberWsTo;
 import com.unimuenster.govlearnapi.group.controller.wsto.GetGroupsWsTo;
 import org.junit.jupiter.api.BeforeEach;
